services/event-consumer-leaderboard/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.0
